requirements.txt
gradio
ultralytics
opencv-python
numpy
Pillow
torch
